requirements.txt
streamlit>=1.28.0
rdkit>=2022.9.5
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.6.0
pillow>=9.5.0
networkx>=3.0